seta-ui/requirements.txt
Werkzeug==2.3.8
flask==2.3.3
flask-cors==4.0.0
pymongo==4.6.0
pycryptodome==3.19.1
python-cas==1.6.0
gunicorn==21.2.0
flask_jwt_extended==4.5.3
flask_apscheduler==1.13.1
flask-pymongo==2.3.0
flask-restx==1.2.0
injector==0.21.0
python-interface==1.6.1
flask-injector==0.15.0
github-flask==3.2.0
requests==2.31.0
shortuuid==1.0.11
pytest==7.4.3
mongodb-migrations==1.3.0
configparser==6.0.0
email-validator==2.1.0.post1
pydantic==2.5.2

psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5